+++ KiLib/Raster/Raster.test.cpp
@@ -61,7 +61,7 @@ namespace KiLib
 
       fs::current_path(path);
 
-      for (auto it : fs::directory_iterator(path))
+      for (const auto &it : fs::directory_iterator(path))
       {
          std::string base = (path / fs::path(it.path().stem().string())).string();
          files.insert(std::make_pair<std::string, std::string>(base + ".tif", base + ".dem"));
@@ -82,7 +82,7 @@ namespace KiLib
 
       fs::current_path(path);
 
-      for (auto it : fs::directory_iterator(path))
+      for (const auto &it : fs::directory_iterator(path))
       {
          if (!it.is_regular_file() && it.path().extension() != ".tif")
             continue;
@@ -113,7 +113,7 @@ namespace KiLib
 
       fs::current_path(path);
 
-      for (auto it : fs::directory_iterator(path))
+      for (const auto &it : fs::directory_iterator(path))
       {
          if (!it.is_regular_file() && it.path().extension() != ".tif")
             continue;
@@ -144,7 +144,7 @@ namespace KiLib
 
       fs::current_path(path);
 
-      for (auto it : fs::directory_iterator(path))
+      for (const auto &it : fs::directory_iterator(path))
       {
          if (!it.is_regular_file() && it.path().extension() != ".tif")
             continue;
@@ -175,7 +175,7 @@ namespace KiLib
 
       fs::current_path(path);
 
-      for (auto it : fs::directory_iterator(path))
+      for (const auto &it : fs::directory_iterator(path))
       {
          if (!it.is_regular_file() && it.path().extension() != ".tif")
             continue;
@@ -207,7 +207,7 @@ namespace KiLib
 
       fs::current_path(path);
 
-      for (std::string size : sizes)
+      for (const std::string &size : sizes)
       {
          std::string base = (path / size).string();
 
@@ -219,4 +219,104 @@ namespace KiLib
       }
    }
 
+   TEST(Raster, getCellPos)
+   {
+      auto                     cwd  = fs::current_path();
+      auto                     path = fs::path(std::string(TEST_DIRECTORY) + "/ComputeSlope/");
+      std::vector<std::string> sizes{"5x5", "7x7", "7x3"};
+
+      fs::current_path(path);
+
+      for (const std::string &size : sizes)
+      {
+         std::string base = (path / size).string();
+
+         Raster dem(base + ".dem");
+         for (size_t r = 0; r < dem.nRows; r++)
+         {
+            for (size_t c = 0; c < dem.nCols; c++)
+            {
+               size_t      ind = dem.flattenIndex(r, c);
+               KiLib::Vec3 pos = dem.getCellPos(ind);
+               double      z   = dem.at(r, c);
+
+               ASSERT_DOUBLE_EQ(pos.z, z);
+               ASSERT_DOUBLE_EQ(pos.y, dem.yllcorner + (double)r * dem.cellsize);
+               ASSERT_DOUBLE_EQ(pos.x, dem.xllcorner + (double)c * dem.cellsize);
+               ASSERT_EQ(ind, dem.getNearestCell(pos));
+            }
+         }
+      }
+   }
+
+   TEST(Raster, Rasterize)
+   {
+      class TestClass
+      {
+      public:
+         TestClass(KiLib::Vec3 pos, double safetyFactor)
+         {
+            this->pos          = pos;
+            this->safetyFactor = safetyFactor;
+         }
+
+         KiLib::Vec3 pos;
+         double      safetyFactor;
+      };
+
+      auto                   cwd  = fs::current_path();
+      auto                   path = fs::path(std::string(TEST_DIRECTORY) + "/ComputeSlope/7x3_NODATA.dem");
+      Raster                 dem(path.string());
+      std::vector<TestClass> objs{
+         {{0.4, 0.3, -1.0}, 0.03},  // 0, 0
+         {{0.1, 0.49, -1.0}, 0.12}, // 0, 0
+         {{0.5, 0.5, -1.0}, 0.3},   // 1, 1
+         {{1.2, 5.1, -1.0}, 0.7},   // 5, 1
+         {{0.9, 4.6, -1.0}, 0.3},   // 5, 1
+         {{1.4, 4.6, -1.0}, 1.1},   // 5, 1
+      };
+
+      //clang-format off
+      KiLib::Raster rasterizedNoTrunc = KiLib::Raster::Rasterize<TestClass>(
+         dem,                                                  // Reference dem to construct sizes
+         objs,                                                 // Our vector of objects to rasterize
+         [](const TestClass &obj) { return obj.pos; },         // Our objects position
+         [](const TestClass &obj) { return obj.safetyFactor; } // Our objects attribute to rasterize
+      );
+      //clang-format on
+
+      ASSERT_DOUBLE_EQ(rasterizedNoTrunc(5, 1), 0.70);
+      ASSERT_DOUBLE_EQ(rasterizedNoTrunc(1, 1), 0.30);
+      ASSERT_DOUBLE_EQ(rasterizedNoTrunc(0, 0), 0.075);
+
+      //clang-format off
+      KiLib::Raster rasterizedTrunc = KiLib::Raster::Rasterize<TestClass>(
+         dem,                                                        // Reference dem to construct sizes
+         objs,                                                       // Our vector of objects to rasterize
+         [](const TestClass &obj) { return obj.pos; },               // Our objects position
+         [](const TestClass &obj) { return obj.safetyFactor < 1.0; } // Our objects attribute to rasterize
+      );
+      //clang-format on
+
+      ASSERT_DOUBLE_EQ(rasterizedTrunc(5, 1), 2.0 / 3.0);
+      ASSERT_DOUBLE_EQ(rasterizedTrunc(1, 1), 1.0);
+      ASSERT_DOUBLE_EQ(rasterizedTrunc(0, 0), 1.0);
+   }
+
+
+   TEST(Raster, GetAverage)
+   {
+      auto   cwd  = fs::current_path();
+      auto   path = fs::path(std::string(TEST_DIRECTORY) + "/ComputeSlope/7x3_NODATA.dem");
+      Raster dem(path.string());
+
+      double avg1 = dem.GetAverage(0, 4.0);
+      ASSERT_DOUBLE_EQ(avg1, 6.0);
+
+      double avg2 = dem.GetAverage(4, 1.0);
+      ASSERT_DOUBLE_EQ(avg2, 4.0);
+
+      double avg3 = dem.GetAverage(14, 2.0);
+      ASSERT_DOUBLE_EQ(avg3, 13.5);
+   }
 } // namespace KiLib