+++ CHANGELOG.md
@@ -2,7 +2,21 @@
 
 Notable changes to this project will be documented in this file.
 
+## vX.X.X
+---
+### Added
+
+- Add flattenIndex method to Raster
+- Add Rasterize method to raster
+- Add GetAverage method to raster
+
+### Fixed
+
+- When compiling on windows we define _USE_MATH_DEFINES to make sure there are no math constant issues
+- getCellPos div by zero
+
 ## v2.5.0 (2020-03-03)
+---
 
 ### Added